+++ src/main/java/calculator/OutputView.java
@@ -4,7 +4,7 @@ public class OutputView {
 
     private static final String RESULT_MESSAGE = "결과 : ";
 
-    public void printResult(Integer sum) {
+    public void printResult(Double sum) {
         System.out.println(RESULT_MESSAGE + sum);
     }
 }
